refactor( 리뷰반영 ) : 테스트코드 위치를 도메인과 같은 단계로 변경
